+++ packages/store/src/index.js
@@ -1,0 +1,6 @@
+export * from './store';
+export * from './ledger';
+export * from './weakStore';
+
+// @deprecated use named export
+export { makeStore as default } from './store';