+++ packages/weak-store/src/index.js
@@ -1,0 +1,2 @@
+// @deprecated Use named "weakStore" export of "@agoric/store"
+export { weakStore as default } from '@agoric/store';