+++ packages/zoe/src/contractFacet/contractFacet.js
@@ -9,7 +9,7 @@
 
 import { assert, details } from '@agoric/assert';
 import { E } from '@agoric/eventual-send';
-import makeWeakStore from '@agoric/weak-store';
+import { makeWeakStore } from '@agoric/store';
 
 import { makeAmountMath, MathKind } from '@agoric/ertp';
 import { makeNotifierKit, updateFromNotifier } from '@agoric/notifier';