+++ packages/weak-store/NEWS.md
@@ -1,5 +1,10 @@
 User-visible changes in @agoric/weak-store:
 
+## Next release
+
+This package as a whole is deprecated. Use the explicitly named
+`makeWeakStore` export of `@agoric/store` instead.
+
 ## Release 0.0.1 (3-Feb-2020)
 
 Moved out of ERTP and created new npm package `@agoric/weak-store`.